requirements.txt
abstract_singleton==1.0.1
accelerate==0.18.0
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
argilla==1.6.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==23.1.0
auto_gpt_plugin_template==0.0.3
autoflake==2.1.1
backoff==2.2.1
beautifulsoup4==4.12.2
bitsandbytes==0.38.1
black==23.3.0
blinker==1.6.2
blis==0.7.9
cachetools==5.3.0
captcha-solver==0.1.5
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.21
click==8.1.3
clickhouse-connect==0.5.22
colorama==0.4.6
commonmark==0.9.1
confection==0.0.4
contourpy==1.0.7
coverage==7.2.5
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.7
datasets==2.12.0
Deprecated==1.2.13
dill==0.3.6
discord==2.2.3
discord.py==2.2.3
diskcache==5.6.1
distlib==0.3.6
distro==1.8.0
dnspython==2.3.0
docker==6.0.1
duckdb==0.7.1
duckduckgo-search==2.9.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fastapi==0.95.1
ffmpy==0.3.0
filelock==3.12.0
flake8==6.0.0
Flask==2.3.2
flask-cloudflared==0.0.12
flexgen==0.1.7
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
ghp-import==2.1.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
GoogleBard==1.0.0
gradio==3.25.0
gradio_client==0.1.4
greenlet==2.0.1
gTTS==2.3.1
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httplib2==0.22.0
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.14.1
identify==2.5.23
idna==3.4
iniconfig==2.0.0
InstructorEmbedding==1.0.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
langchain==0.0.162
langcodes==3.3.0
linkify-it-py==2.0.2
llama-cpp-python==0.1.41
loguru==0.7.0
lxml==4.9.2
lz4==4.3.2
Markdown==3.3.7
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mccabe==0.7.0
mdit-py-plugins==0.3.3
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.4.2
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
nodeenv==1.7.0
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
olefile==0.46
openai==0.27.2
openapi-python-client==0.13.4
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.8.10
outcome==1.2.0
packaging==23.1
pandas==1.5.3
pathspec==0.11.1
pathy==0.10.1
pdf2image==1.16.3
pdfminer.six==20221105
peft @ git+https://github.com/huggingface/peft@a37156c2c7966ff7d1487d33734738414aa0bafc
Pillow==9.5.0
pinecone-client==2.2.1
platformdirs==3.5.0
playsound==1.2.2
playwright==1.32.1
pluggy==1.0.0
posthog==3.0.1
pre-commit==3.3.0
preshed==3.0.8
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
PuLP==2.7.0
py-cpuinfo==9.0.0
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.7
pydub==0.25.1
pyee==9.0.4
pyflakes==3.0.1
PyGithub==1.58.1
Pygments==2.15.1
PyJWT==2.6.0
pyllamacpp==1.0.7
pymdown-extensions==9.11
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pypandoc==1.11
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-benchmark==4.0.0
pytest-cov==4.0.0
pytest-integration==0.2.3
pytest-mock==3.10.0
pytest-recording==0.12.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-git==2018.2.1
python-http-client==3.3.7
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
pyyaml_env_tag==0.1
rank-bm25==0.2.2
readability-lxml==0.8.1
redis==4.5.4
regex==2023.3.23
requests==2.29.0
requests-oauthlib==1.3.1
responses==0.18.0
rfc3986==1.5.0
rich==13.0.1
rsa==4.9
rwkv==0.7.3
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.1.4
semantic-version==2.10.0
Send2Trash==1.8.2
sendgrid==6.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.0.post1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.12
srsly==2.4.6
starkbank-ecdsa==2.2.0
starlette==0.26.1
sympy==1.11.1
tenacity==8.2.2
thinc==8.1.9
threadpoolctl==3.1.0
tiktoken==0.3.3
tokenizers==0.13.3
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.28.1
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.14.0
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.2
undetected-chromedriver==3.1.7
unstructured==0.6.4
uritemplate==4.1.1
urllib3==1.26.15
urllib3-secure-extra==0.1.0
uvicorn==0.22.0
uvloop==0.17.0
vcrpy==4.2.1
virtualenv==20.23.0
wasabi==1.1.1
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.6
webdriver-manager==3.8.6
websocket-client==1.5.1
websockets==11.0.2
Werkzeug==2.3.3
wrapt==1.14.1
wsproto==1.2.0
XlsxWriter==3.1.0
xxhash==3.2.0
yarl==1.9.2
zstandard==0.21.0
